perf(home): defer js code & move code to server side
