+++ components/Link.tsx
@@ -6,7 +6,9 @@ import type { LinkProps } from 'next/link'
 import { AnchorHTMLAttributes, useEffect, useState } from 'react'
 import { useSearchParams } from 'next/navigation'
 
-const CustomLink = ({ href, ...rest }: LinkProps & AnchorHTMLAttributes<HTMLAnchorElement>) => {
+type CustomLinkProps = LinkProps & AnchorHTMLAttributes<HTMLAnchorElement>
+
+const CustomLink = ({ href, prefetch, ...rest }: CustomLinkProps) => {
   const searchParams = useSearchParams()
   const [regionParam, setRegionParam] = useState<string | null>(null)
   const [cloudRegionParam, setCloudRegionParam] = useState<string | null>(null)
@@ -34,14 +36,14 @@ const CustomLink = ({ href, ...rest }: LinkProps & AnchorHTMLAttributes<HTMLAnch
         newHref = `${newHref}&cloud_region=${cloudRegionParam}`
       }
 
-      return <Link href={newHref} {...rest} target="_blank" prefetch={false} />
+      return <Link href={newHref} {...rest} target="_blank" prefetch={prefetch ?? false} />
     }
 
     if (href.startsWith('https://signoz.io/')) {
-      return <Link href={href} {...rest} target="_blank" prefetch={false} />
+      return <Link href={href} {...rest} target="_blank" prefetch={prefetch ?? false} />
     }
 
-    return <Link href={href} {...rest} prefetch={false} />
+    return <Link href={href} {...rest} prefetch={prefetch ?? false} />
   }
 
   if (isAnchorLink) {
