+++ components/observability/index.tsx
@@ -1,4 +1,3 @@
-'use client'
 import React from 'react'
 import Image from 'next/image'
 import Button from '../../components/ui/Button'
