+++ utils/userClient.ts
@@ -3,9 +3,6 @@ import { v4 as uuidv4 } from 'uuid'
 const ANONYMOUS_ID_KEY = 'app_anonymous_id'
 const USER_ID_KEY = 'app_user_id'
 
-/**
- * Get or create anonymous ID for the user (client-side only)
- */
 export const getOrCreateAnonymousId = (): string | undefined => {
   if (typeof window === 'undefined') return undefined
 
@@ -18,29 +15,17 @@ export const getOrCreateAnonymousId = (): string | undefined => {
     }
 
     return id || undefined
-  } catch (error) {
+  } catch {
     return undefined
   }
 }
 
-/**
- * Get user ID from localStorage (client-side only)
- */
 export const getUserId = (): string | undefined => {
   if (typeof window === 'undefined') return undefined
 
   try {
     return localStorage.getItem(USER_ID_KEY) || undefined
-  } catch (error) {
+  } catch {
     return undefined
   }
-}
-
-/**
- * Extract group ID from email
- */
-export const extractGroupIdFromEmail = (email?: string): string | undefined => {
-  if (!email) return undefined
-  const parts = email.split('@')
-  return parts.length === 2 ? parts[1] : undefined
 }
